0.3.1 (#20)
* 0.3.1

Update versions and change CSS classes in components"

Upgraded the project version from 0.3.0 to 0.3.1 and updated several library versions in package.json and package-lock.json. Also changed the CSS class names in DocsCard.tsx for better consistency with the app's color scheme and design. These updates ensure that the project dependencies stay up-to-date and the app maintains a coherent look and feel.

* Improve GitHub CI workflow and testing

Modified commands to start Supabase in both the CI workflow and test script and introduced artifact archiving for test screenshots.

The 'supabase start' command was updated to exclude unnecessary services, to speed up testing and reduce resource usage. Added an additional step to archive screenshots generated from Cypress end-to-end testing in the case of failure, increasing visibility into failed test cases. Artifacts are kept for 2 days for review.