0.4.9 (#32)
* Reduce DashboardDemo data range, update package version and add runtime variable

The range of data generated for the DashboardDemo component was reduced from 0-1000 to 0-10 as a part of performance improvement. The version number of the package was updated in package.json and package-lock.json to reflect recent changes. A "runtime" constant was added in layout.tsx files to manage the environment in a better way.

* Use edge env only in auth and site layouts, where it's less likely developers will add Node.js modules

* Improve accessibility and style consistency in navigation elements

This commit makes several updates in our components to improve both style and accessibility. Firstly changes have been made to our Footer so that it uses <span> elements instead of <Heading> for the "About", "Product", and "Legal" sections. This makes our codebase more consistent as well as better matching the actual visual style of the page. In our NavigationItem, the link elements have been wrapped in <li> elements to improve overall semantic value. Finally, an 'aria-label' was added to our DropdownMenuTrigger in the MobileDropdown function to aid with accessibility.

* Simplify NewsletterSignup style and remove unused `runtime` variables

The header style in `NewsletterSignup.tsx` was simplified to remove unnecessary complexity. Instead of the Heading component, a simple span with class style was used for the header text. Removed the `runtime` variable from both `layout.tsx` in the main site and auth sections, since it was not used anywhere.

* Refactor UserSession retrieval for site layout

User session data retrieval was refactored from an asynchronous function calling loadUserData() in SiteLayout component to useQuery hook inside the new useUserSession(). This change makes use of the more modern and efficient SWR library for data fetching. It will improve data fetching efficiency and reduce repetitive code as useQuery handles caching, revalidation and errors out of the box. This update also tickets off the need for manual session data handling on every render.

In addition to this, props.data was replaced with data in SiteHeaderSessionProvider component. useEffect hook was implemented to handle setting user session data whenever "data" prop changes improving, in this way, rendering efficiency.

* Try edge runtime in auth

* Refactor session handling in layout and components

The commit removed session fetching logic from the layout file and embedded the same into the SiteHeaderSessionProvider component. This was done to centralize and simplify the fetching and handling of the user session. Now, whenever a user session is required, the SiteHeaderSessionProvider is used, which internally manages the session and allows its children to access it through an accessible context. This eliminates redundant fetch calls and makes components using the user session much simpler.

* Improve data loading and refactor layout components

The data loading process was refactored in server/loaders/load-user-data.ts to efficiently handle user data and user sessions. The `language` attribute in the configuration.ts file was removed as it was no longer necessary. The loading.tsx component was updated to include a more user-friendly loading indicator called `TopLoadingBarIndicator`. The SiteLayout function in layout.tsx was changed to an async method, better handling data fetching for user sessions, which also led to changes in the `SiteHeaderSessionProvider.tsx` where querying for user session data involves less code now, simplifying it. In `PricingTable.tsx`, the key was changed from `stripePriceId` to `plan.name` to ensure uniqueness even when `stripePriceId` does not exist because plan names are unique. The changes were made to improve the user experience and refactor the process to load user data.

* Change import declaration and update CSS class

Updated import declaration in Post.tsx to import 'Post' as a type for a more explicit codebase. Improved visual readability in not-found.tsx by adding a class to allow better differentiation in dark mode.

* Refactor NavigationItem component for cleaner code

Shifted className property from Link element to its parent li element in NavigationItem.tsx for simpler structure. Adjusted certain classNames to simplify and clean-up the code, leading to a more readable and manageable codebase. Also, removed redundant transition-colors to avoid any potential bugs or inconsistencies in future.