0.4.11 (#34)
Improve initialization process of Sentry.

This commit rearranges the initialization process of Sentry for performance and code flow logic. It moves the import of `@sentry/react` to a location where it's only invoked if a DSN is present. This reduces the overall load when a DSN is not provided and allows a smoother operation in non-production environments.
Another change is that it avoids calling `warnSentryNotConfigured()` when in non-production environment and DSN isn't available. However, if DSN is not available in production, a warning message is displayed and the function is returned. This ensures error tracking is only enabled when the Sentry DSN environment variable is present.