Add authentication error handling page and update package dependencies (#40)
A new Error Page has been created under 'src/app/auth/callback/error' for handling authentication errors. This page displays a detailed error message if there's an error during the authentication callback and provides a way for users to resend the authentication link. The error page is designed to redirect users back to the sign-in page in case there's no error.

Along with this, multiple packages in package-lock.json were updated to their latest versions. This includes updates to sentry, stripe and pino to name a few. These updates do not contain any breaking changes and were made in part due to important bug fixes, performance improvements and new features being added in the newer versions.