Update the version and dependencies in package-lock file (#41)
* Update the version and dependencies in package-lock file

Updated the package-lock.json file to reflect the new app version 0.5.5 from 0.5.4, updated dependencies, and moved from experimental to final APIs for useFormState and useFormStatus. Also added legacy-peer-deps=true to .npmrc for backward compatibility.

These changes are needed to keep the application packages up-to-date and to move out of the experimental APIs in useFormState and useFormStatus which may cause instability in the application.

* Updated to Supabase SSR packages, deprecated Auth Helpers

* Add Email OTP authentication

Added the option for Email OTP (One-Time Password) authentication. A new container 'EmailOtpContainer.tsx' was created containing functions to verify the OTP code. This option was added to both SignInMethodsContainer.tsx and SignUpMethodsContainer.tsx. The option was also included in configuration.ts to allow straightforward enablement or disablement of this authentication method.

This feature provides an additional layer of security and offers users an alternative method for authentication.

* Refactor authentication for testing and server loading

The authentication process for both the server loading and testing logic has been refactored. This aims to improve maintainability, readability and to ensure user information is validated correctly. Specifically, for server load-app-data, we have removed unnecessary return statements for redirectToHomePage and redirect functions as they were not being used as promises. In testing commands, we ensured programmatic sign-ins return a response that is validated to check if the user is authenticated. Moreover, for auth.po in testing, we replaced the older more convoluted approach to sign users in programmatically with the createBrowserClient utility from supabase. This makes the tests more robust and reliable. Also, the validation method was added to verify and return user data for testing purposes.