0.5.6 (#42)
Replace Supabase client references to improve code clarity

Renamed getSupabaseServerClient to getSupabaseServerComponentClient and getSupabaseServerActionClient to getSupabaseRouteHandlerClient. These changes were made to better reflect the clients' roles in handling actions on the server and routing, respectively. This will make our codebase more easily navigable and understandable for any developers working on it in the future.