0.5.8 (#44)
* 0.5.8

Remove supabase cookie adapter and bump package versions

The `supabase-cookie-adapter.ts` has been removed since it is no longer necessary due to the adoption of a more simplified cookie strategy. Adjustments were made in `action-client.ts` to reflect this change and simplify the use of cookies. Meanwhile, the project package versions were updated accordingly in `package-lock.json`.

* Update @supabase/ssr from version 0.0.6 to 0.0.7

In this commit, we have updated the version of @supabase/ssr from 0.0.6 to 0.0.7.