Update package version and enhance UI interactions (#50)
* Update package version and enhance UI interactions

Increased the package version from 0.5.14 to 0.5.15 and updated the 'supabase' dependency version for improvements. Additionally, modified NavigationItem for better handling of disabled links and updated ProfileSettingsTabs to disable edit on Email link if password editing is not allowed.

* Remove CSRF token usage in various components and actions

Eliminated the usage of CSRF token in several user and profile action components. The removal encompassed fetch hooks, hidden form inputs, and function parameters in user action files. Also, the changes have been extended to the stripe actions library and in the middleware file.

* Update non-admin user test URL assertion

Updated the URL assertion in the non-admin user end-to-end test. Previously, the assertion expected the URL to be the homepage address. Now, the test correctly asserts a redirection to the 404 error page.

* Change site URL environment variable name

The name of the site URL environment variable has been changed from 'SITE_URL' to 'NEXT_PUBLIC_SITE_URL'. Also, the expected URL in the non-admin user test has been updated to align with this change. Now, both places use the local IP address instead of 'localhost'.

* Remove Sentry Tracing and update npm configuration.

The "@sentry/tracing" dependency was removed from the Sentry initialization file and the "package.json" file, as it isn't required anymore. The tracing is now carried out via "@sentry/browser" package. A "shamefully-hoist" option was also added to the ".npmrc" file to alter the package installation behavior.