0.6.0 (#70)
Refactor UpdateProfileForm and upgrade packages

The UpdateProfileForm was refactored to streamline the functionality and remove unnecessary caching. This change simplifies the process of updating user profile information. In addition, several packages needed for the application were upgraded to their latest versions. This ensures that the application has the latest features and security updates available.